igniteera-backend/igniteera/requirements.txt
Django
djangorestframework 
djangorestframework-simplejwt 
django-cors-headers 
python-dotenv 
